Remove all primary keys from seed files
The primary keys are all auto-incrementing and the database will add them automatically and by removing them we avoid having to update a long row of keys if we insert a record in the middle of the list. Further, we do not have to think about when the port waypoints list stopped when we are seeding suggested route waypoints.