Fix waypoint constraints (#64)
* Make port ID optional

* Make suggeste route optional

* Print seed file names to console

* Seed port waypoints before suggested route waypoints

* Replace dates with datetime

* Point voyages to port waypoints

* IDs on suggested route waypoints should start after port waypoints end

* Remove ports from suggested route waypoints

* Do not empty table before running this script

This table will cleared by the port waypoints seed. Clearing the table now will remove the data added by the port waypoint seed.

* Remove all primary keys from seed files

The primary keys are all auto-incrementing and the database will add them automatically and by removing them we avoid having to update a long row of keys if we insert a record in the middle of the list. Further, we do not have to think about when the port waypoints list stopped when we are seeding suggested route waypoints.

* Add sequence IDs to suggested route waypoints

Since all of the waypoints belong to different suggested routes, they all start at 1.